requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.11.0
asyncio==3.4.3
bcrypt==5.0.0
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
coverage==7.13.1
cycler==0.12.1
dj-database-url==3.1.2
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dominate==2.9.1
fastapi==0.121.3
Flask==3.1.2
flask-cors==6.0.1
Flask-Mail==0.10.0
fonttools==4.61.1
greenlet==3.3.1
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
multidict==6.7.1
nltk==3.9.2
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
propcache==0.4.1
pycparser==3.0
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
RapidFuzz==3.14.3
regex==2025.11.3
requests==2.32.5
rich==14.2.0
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.4
starlette==0.50.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
unicorn==2.1.4
urllib3==2.5.0
uvicorn==0.38.0
waitress==3.0.2
Werkzeug==3.1.3
whitenoise==6.11.0
